Do lint fixes

+++ AdaptFM/model/foundation_models/cellposeSAM/inference_wrapper.py
@@ -1,4 +1,5 @@
 import argparse
+from pathlib import Path
 from cellpose import models 
 import os
 import tifffile as tiff
@@ -27,15 +28,18 @@ def main():
         model = models.CellposeModel(gpu=True)
     
     # run model on test images
-    for image_name in os.listdir(test_dir):
-
-        image_path = os.path.join(test_dir,image_name)
+    for image_obj in Path(test_dir).iterdir():
+        if not image_obj.is_file():
+            continue
+            
+        image_name = image_obj.name
+        image_path = str(image_obj)
         image = tiff.imread(image_path)
 
-        masks, _, _ = model.eval(image,do_3D=True,channel_axis=3,z_axis=0)
+        masks, _, _ = model.eval(image, do_3D=True, channel_axis=3, z_axis=0)
 
-        image_output_path = os.path.join(output_path,image_name)
-        tiff.imwrite(image_output_path,masks)
+        image_output_path = str(Path(output_path) / image_name)
+        tiff.imwrite(image_output_path, masks)
 
 if __name__ == "__main__":
     main()
